refactor(stores): add document identity and injected store factories
Steps 1-3 of the #53 document registry plan. No registry yet and no
user-visible change: the app still hosts exactly one implicit document.

1. Document identity primitives. `DocumentId` is a branded string so a
   file path can never be passed where an identity is required, and
   `createDocumentId()` draws from `crypto.randomUUID`/`getRandomValues`
   and throws rather than falling back to `Math.random()`.

2. Mounted-canvas state moved out of the canvas store into the new
   workspace-scoped `canvas-instance-store`: the ReactFlow handles
   (`rfFitView`/`rfZoomIn`/`rfZoomOut`/`rfPanBy`), the palette `dragged*`
   fields, `isConnecting`, and `altDragActive`. `setDraggedType` is
   dropped; it had no caller. `onNodesChange` keeps one intentional
   cross-store read of `altDragActive`, now covered by a test.

3. Model, canvas, and history stores are produced by `create*Store(deps)`
   factories. The canvas factory receives its model and history stores by
   injection instead of importing the globals, so a canvas store cannot
   write into another document. Every module-level mutable moved into a
   factory closure: the history-capture debounce, `preDragSnapshot`,
   `nudgeState`, and the three id counters. `model-store.ts`,
   `canvas-store.ts`, and `history-store.ts` are now facades over the
   active bundle held in `document-stores.ts`, preserving the
   `useXStore(selector)` / `.getState()` / `.setState()` shape and the
   pure-helper exports. `clipboard-store.ts` is the only consumer that
   changed; it allocates ids through the canvas store actions.

Refs #53
